requirement.txt
torch
pillow
tensorflow
matplotlib
torchvision
PyYAML
uvicorn
fastapi
opencv-python
numpy
starlette
google
